moved buttons to hamburger menu
and color stuff
