+++ client/app/components/Card.js
@@ -19,7 +19,6 @@ export default class Card extends React.Component {
 	}
 
 	_onClickHandle(e) {
-		console.log('Data to display in next screen: ', this.props.data);
 		this.props.navigation.navigate('RequestDetailsScreen', {data: this.props.data});
 	}
